Separate intersection.js
