+++ packages/turf-line-offset/index.js
@@ -1,6 +1,7 @@
+var helpers = require('@turf/helpers');
 var getCoords = require('@turf/invariant').getCoords;
 var coordEach = require('@turf/meta').coordEach;
-var helpers = require('@turf/helpers');
+var intersection = require('./intersection');
 var lineString = helpers.lineString;
 var distanceToDegrees = helpers.distanceToDegrees;
 
@@ -15,16 +16,19 @@ var distanceToDegrees = helpers.distanceToDegrees;
  * @example
  * var line = {
  *   "type": "Feature",
- *   "properties": {},
+ *   "properties": {
+ *     "stroke": "#F00"
+ *   },
  *   "geometry": {
  *     "type": "LineString",
  *     "coordinates": [[-83, 30], [-84, 36], [-78, 41]]
  *   }
  * };
  *
- * var offsetLine = turf.lineOffset(line, 2, 'miles');
+ * var offsetLine = turf.lineOffset(line, 2, "miles");
  *
  * //addToMap
+ * offsetLine.properties.stroke = "#00F"
  * var addToMap = [offsetLine, line]
  */
 module.exports = function (line, offset, units) {
@@ -83,115 +87,4 @@ function processSegment(point1, point2, offset) {
     var out1y = point1[1] + offset * (point1[0] - point2[0]) / L;
     var out2y = point2[1] + offset * (point1[0] - point2[0]) / L;
     return [[out1x, out1y], [out2x, out2y]];
-}
-
-/**
- * AB
- *
- * @private
- * @param {Array<Array<number>>} segment - 2 vertex line segment
- * @returns {Array<number>} coordinates [x, y]
- */
-function ab(segment) {
-    var start = segment[0];
-    var end = segment[1];
-    return [end[0] - start[0], end[1] - start[1]];
-}
-
-/**
- * Cross Product
- *
- * @private
- * @param {Array<number>} v1 coordinates [x, y]
- * @param {Array<number>} v2 coordinates [x, y]
- * @returns {Array<number>} Cross Product
- */
-function crossProduct(v1, v2) {
-    return (v1[0] * v2[1]) - (v2[0] * v1[1]);
-}
-
-/**
- * Add
- *
- * @private
- * @param {Array<number>} v1 coordinates [x, y]
- * @param {Array<number>} v2 coordinates [x, y]
- * @returns {Array<number>} Add
- */
-function add(v1, v2) {
-    return [v1[0] + v2[0], v1[1] + v2[1]];
-}
-
-/**
- * Sub
- *
- * @private
- * @param {Array<number>} v1 coordinates [x, y]
- * @param {Array<number>} v2 coordinates [x, y]
- * @returns {Array<number>} Sub
- */
-function sub(v1, v2) {
-    return [v1[0] - v2[0], v1[1] - v2[1]];
-}
-
-/**
- * scalarMult
- *
- * @private
- * @param {number} s scalar
- * @param {Array<number>} v coordinates [x, y]
- * @returns {Array<number>} scalarMult
- */
-function scalarMult(s, v) {
-    return [s * v[0], s * v[1]];
-}
-
-/**
- * Intersect Segments
- *
- * @private
- * @param {Array<number>} a coordinates [x, y]
- * @param {Array<number>} b coordinates [x, y]
- * @returns {Array<number>} intersection
- */
-function intersectSegments(a, b) {
-    var p = a[0];
-    var r = ab(a);
-    var q = b[0];
-    var s = ab(b);
-
-    var cross = crossProduct(r, s);
-    var qmp = sub(q, p);
-    var numerator = crossProduct(qmp, s);
-    var t = numerator / cross;
-    var intersection = add(p, scalarMult(t, r));
-    return intersection;
-}
-
-/**
- * Is Parallel
- *
- * @private
- * @param {Array<number>} a coordinates [x, y]
- * @param {Array<number>} b coordinates [x, y]
- * @returns {boolean} true if a and b are parallel (or co-linear)
- */
-function isParallel(a, b) {
-    var r = ab(a);
-    var s = ab(b);
-    return (crossProduct(r, s) === 0);
-}
-
-/**
- * Intersection
- * https://github.com/rook2pawn/node-intersection/blob/master/index.js
- *
- * @private
- * @param {Array<number>} a coordinates [x, y]
- * @param {Array<number>} b coordinates [x, y]
- * @returns {Array<number>|boolean} true if a and b are parallel (or co-linear)
- */
-function intersection(a, b) {
-    if (isParallel(a, b)) return false;
-    return intersectSegments(a, b);
 }
